Fix for CC to non system user
In situations where a CC'd user was added but that email address is not
associated with an active system user, a 500 error would be thrown.
Added a straight-forward rescue block in that situation. If you have
filtered enabled on the global pgp settings, this doesnt present a risk
since that user will merely be thrown a notice that they need to login to
set their pgp key to see future messages.