Revert "Encrypted password reset e-mails"
This reverts commit 24931b88634e955d2456acc9311ff63edc03f951. The reasoning
behind the revert is that if a user doesnt take advantage of the PGP feature
and actually upload their public key, they are unable to self-reset without
admin intervention. This is problematic and until we enforce public keys
during the account spin-up process we are going to have to lax this enforcement
to make the site more realistic to manage at scale.